requirements.txt
astroid==2.2.5
Click==7.0
Flask==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.3.1
python-dotenv==0.10.3
six==1.12.0
SQLAlchemy==1.3.5
typed-ast==1.4.0
Werkzeug==0.15.5
wrapt==1.11.2
